+++ .github/FUNDING.yml
@@ -1,7 +1,7 @@
 # These are supported funding model platforms
 
 github: # Replace with up to 4 GitHub Sponsors-enabled usernames e.g., [user1, user2]
-patreon: conrevo # Replace with a single Patreon username
+patreon: # Replace with a single Patreon username
 open_collective: # Replace with a single Open Collective username
 ko_fi: conrevo # Replace with a single Ko-fi username
 tidelift: # Replace with a single Tidelift platform-name/package-name e.g., npm/babel
